Decompile NP3 EntityGaibon
