+++ config/symbols.us.stnp3.txt
@@ -61,26 +61,13 @@ EntitySlograSpearProjectile = 0x801B7798;
 EntityGaibon = 0x801B78BC;
 EntitySmallGaibonProjectile = 0x801B8D84;
 EntityLargeGaibonProjectile = 0x801B8E94;
-Random = 0x801B90BC;
 Update = 0x801B90EC;
 UpdateStageEntities = 0x801B93E8;
 HitDetection = 0x801B94F0;
 EntityDamageDisplay = 0x801BA608;
 InitRoomEntities = 0x801BB508;
 UpdateRoomPosition = 0x801BB680;
-CreateEntityFromCurrentEntity = 0x801BB734;
-CreateEntityFromEntity = 0x801BB7A8;
 EntityRedDoor = 0x801BB89C;
-DestroyEntity = 0x801BC4D4;
-AnimateEntity = 0x801BC604;
-GetDistanceToPlayerX = 0x801BC7D4;
-GetDistanceToPlayerY = 0x801BC810;
-GetSideToPlayer = 0x801BC844;
-MoveEntity = 0x801BC888;
-AllocEntity = 0x801BCD44;
-SetStep = 0x801BD114;
-SetSubStep = 0x801BD134;
-InitializeEntity = 0x801BD1E4;
 EntityDummy = 0x801BD2E0;
 EntityPrizeDrop = 0x801BDEF4;
 EntityExplosion = 0x801BE768;
